Fix imports after renaming arrow.core.test
