+++ arrow-fx-kotlinx-coroutines/src/test/kotlin/arrow/integrations/kotlinx/CoroutinesIntegrationTest.kt
@@ -6,24 +6,23 @@ import arrow.core.extensions.either.eq.eq
 import arrow.core.extensions.eq
 import arrow.core.internal.AtomicRefW
 import arrow.core.right
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.throwable
+import arrow.core.test.laws.equalUnderTheLaw
 import arrow.fx.ForIO
 import arrow.fx.IO
 import arrow.fx.extensions.exitcase.eq.eq
 import arrow.fx.extensions.fx
 import arrow.fx.extensions.io.applicative.applicative
 import arrow.fx.extensions.io.applicativeError.attempt
-import arrow.fx.extensions.io.async.effectMap
-import arrow.fx.handleErrorWith
 import arrow.fx.extensions.io.bracket.onCancel
 import arrow.fx.extensions.io.concurrent.waitFor
 import arrow.fx.fix
+import arrow.fx.handleErrorWith
 import arrow.fx.typeclasses.Duration
 import arrow.fx.typeclasses.ExitCase
 import arrow.fx.typeclasses.milliseconds
 import arrow.fx.typeclasses.seconds
-import arrow.test.UnitSpec
-import arrow.test.generators.throwable
-import arrow.test.laws.equalUnderTheLaw
 import arrow.test.laws.shouldBeEq
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
