+++ arrow-fx-test/src/main/kotlin/arrow/test/laws/Law.kt
@@ -1,9 +1,9 @@
 package arrow.test.laws
 
-import arrow.test.generators.tuple2
-import arrow.test.generators.tuple3
-import arrow.test.generators.tuple4
-import arrow.test.generators.tuple5
+import arrow.core.test.generators.tuple2
+import arrow.core.test.generators.tuple3
+import arrow.core.test.generators.tuple4
+import arrow.core.test.generators.tuple5
 import arrow.typeclasses.Eq
 import io.kotlintest.Matcher
 import io.kotlintest.Result
