+++ arrow-fx-test/src/main/kotlin/arrow/test/laws/MonadThrowLaws.kt
@@ -2,10 +2,12 @@ package arrow.test.laws
 
 import arrow.Kind
 import arrow.core.extensions.eq
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.fatalThrowable
+import arrow.core.test.generators.throwable
+import arrow.core.test.laws.Law
+import arrow.core.test.laws.MonadErrorLaws
 import arrow.test.eq.throwableEq
-import arrow.test.generators.GenK
-import arrow.test.generators.fatalThrowable
-import arrow.test.generators.throwable
 import arrow.typeclasses.Apply
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
