+++ arrow-fx/src/test/kotlin/arrow/fx/EffectsSuspendDSLTests.kt
@@ -4,15 +4,15 @@ import arrow.Kind
 import arrow.core.Left
 import arrow.core.Right
 import arrow.core.Tuple2
-import arrow.core.internal.AtomicIntW
 import arrow.core.identity
+import arrow.core.internal.AtomicIntW
+import arrow.core.test.UnitSpec
 import arrow.fx.extensions.fx
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.io.unsafeRun.runBlocking
 import arrow.fx.extensions.io.unsafeRun.unsafeRun
 import arrow.fx.typeclasses.Concurrent
 import arrow.fx.typeclasses.UnsafeRun
-import arrow.test.UnitSpec
 import arrow.unsafe
 import io.kotlintest.shouldBe
 import io.kotlintest.shouldThrow
