+++ arrow-fx/src/test/kotlin/arrow/fx/EqTest.kt
@@ -1,10 +1,10 @@
 package arrow.fx
 
+import arrow.core.test.UnitSpec
+import arrow.core.test.laws.equalUnderTheLaw
 import arrow.fx.internal.TimeoutException
 import arrow.fx.typeclasses.milliseconds
-import arrow.test.UnitSpec
 import arrow.test.eq.eq
-import arrow.test.laws.equalUnderTheLaw
 import arrow.test.laws.shouldBeEq
 import arrow.test.laws.shouldNotBeEq
 import io.kotlintest.shouldThrow
