+++ arrow-fx/src/test/kotlin/arrow/fx/FiberTest.kt
@@ -1,18 +1,18 @@
 package arrow.fx
 
 import arrow.core.extensions.monoid
+import arrow.core.test.UnitSpec
+import arrow.core.test.laws.ApplicativeLaws
+import arrow.core.test.laws.MonoidLaws
 import arrow.fx.extensions.applicative
 import arrow.fx.extensions.functor
 import arrow.fx.extensions.io.applicative.applicative
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.monoid
 import arrow.fx.typeclasses.Fiber
-import arrow.test.UnitSpec
 import arrow.test.eq.eq
 import arrow.test.eq.eqK
 import arrow.test.generators.genK
-import arrow.test.laws.ApplicativeLaws
-import arrow.test.laws.MonoidLaws
 import io.kotlintest.properties.Gen
 
 class FiberTest : UnitSpec() {
