+++ arrow-fx/src/test/kotlin/arrow/fx/MVarTest.kt
@@ -8,6 +8,8 @@ import arrow.core.Tuple3
 import arrow.core.Tuple4
 import arrow.core.Tuple7
 import arrow.core.extensions.eq
+import arrow.core.test.UnitSpec
+import arrow.core.test.laws.equalUnderTheLaw
 import arrow.core.toT
 import arrow.fx.extensions.fx
 import arrow.fx.extensions.io.async.async
@@ -16,9 +18,7 @@ import arrow.fx.extensions.io.concurrent.parSequence
 import arrow.fx.extensions.io.monad.flatMap
 import arrow.fx.extensions.io.monad.followedBy
 import arrow.fx.typeclasses.milliseconds
-import arrow.test.UnitSpec
 import arrow.test.eq.eq
-import arrow.test.laws.equalUnderTheLaw
 import arrow.test.laws.shouldBeEq
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
