+++ arrow-fx/src/test/kotlin/arrow/fx/QueueTest.kt
@@ -1,21 +1,21 @@
 package arrow.fx
 
+import arrow.core.Left
 import arrow.core.None
 import arrow.core.Some
 import arrow.core.Tuple2
 import arrow.core.Tuple3
-import arrow.core.Left
 import arrow.core.extensions.list.traverse.traverse
 import arrow.core.fix
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.nonEmptyList
+import arrow.core.test.generators.tuple2
+import arrow.core.test.generators.tuple3
 import arrow.fx.extensions.fx
 import arrow.fx.extensions.io.applicative.applicative
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.io.dispatchers.dispatchers
 import arrow.fx.typeclasses.milliseconds
-import arrow.test.UnitSpec
-import arrow.test.generators.nonEmptyList
-import arrow.test.generators.tuple2
-import arrow.test.generators.tuple3
 import io.kotlintest.fail
 import io.kotlintest.matchers.types.shouldBeInstanceOf
 import io.kotlintest.properties.Gen
