+++ arrow-fx/src/test/kotlin/arrow/fx/RefTest.kt
@@ -2,14 +2,14 @@ package arrow.fx
 
 import arrow.Kind
 import arrow.core.extensions.eq
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.functionAToB
+import arrow.core.test.laws.equalUnderTheLaw
 import arrow.fx.extensions.io.concurrent.concurrent
 import arrow.fx.extensions.io.monadDefer.monadDefer
 import arrow.fx.typeclasses.Concurrent
 import arrow.fx.typeclasses.MonadDefer
-import arrow.test.UnitSpec
 import arrow.test.eq.eqK
-import arrow.test.generators.functionAToB
-import arrow.test.laws.equalUnderTheLaw
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
 import io.kotlintest.properties.Gen
