+++ arrow-fx/src/test/kotlin/arrow/fx/ScheduleTest.kt
@@ -11,6 +11,17 @@ import arrow.core.extensions.id.eqK.eqK
 import arrow.core.extensions.id.monad.monad
 import arrow.core.extensions.list.foldable.forAll
 import arrow.core.extensions.monoid
+import arrow.core.test.UnitSpec
+import arrow.core.test.concurrency.SideEffect
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.GenK2
+import arrow.core.test.generators.applicative
+import arrow.core.test.generators.intSmall
+import arrow.core.test.laws.AlternativeLaws
+import arrow.core.test.laws.ApplicativeLaws
+import arrow.core.test.laws.CategoryLaws
+import arrow.core.test.laws.MonoidLaws
+import arrow.core.test.laws.ProfunctorLaws
 import arrow.core.toT
 import arrow.core.value
 import arrow.fx.extensions.io.applicativeError.attempt
@@ -25,17 +36,6 @@ import arrow.fx.extensions.schedule.profunctor.profunctor
 import arrow.fx.typeclasses.Duration
 import arrow.fx.typeclasses.nanoseconds
 import arrow.fx.typeclasses.seconds
-import arrow.test.UnitSpec
-import arrow.test.concurrency.SideEffect
-import arrow.test.generators.GenK
-import arrow.test.generators.GenK2
-import arrow.test.generators.applicative
-import arrow.test.generators.intSmall
-import arrow.test.laws.AlternativeLaws
-import arrow.test.laws.ApplicativeLaws
-import arrow.test.laws.CategoryLaws
-import arrow.test.laws.MonoidLaws
-import arrow.test.laws.ProfunctorLaws
 import arrow.test.laws.forFew
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
