+++ arrow-fx/src/test/kotlin/arrow/fx/data/DurationTest.kt
@@ -1,8 +1,8 @@
 package arrow.fx.data
 
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.intSmall
 import arrow.fx.typeclasses.Duration
-import arrow.test.UnitSpec
-import arrow.test.generators.intSmall
 import arrow.test.generators.timeUnit
 import io.kotlintest.properties.Gen
 import io.kotlintest.properties.forAll
