+++ arrow-streams/src/test/kotlin/arrow/streams/internal/FreeCTest.kt
@@ -31,11 +31,11 @@ import arrow.streams.internal.freec.eq.eq
 import arrow.streams.internal.freec.functor.functor
 import arrow.streams.internal.freec.monad.monad
 import arrow.streams.internal.freec.monadDefer.monadDefer
-import arrow.test.UnitSpec
-import arrow.test.generators.GenK
-import arrow.test.generators.functionAToB
-import arrow.test.generators.throwable
-import arrow.test.laws.EqLaws
+import arrow.core.test.UnitSpec
+import arrow.core.test.generators.GenK
+import arrow.core.test.generators.functionAToB
+import arrow.core.test.generators.throwable
+import arrow.core.test.laws.EqLaws
 import arrow.test.laws.MonadDeferLaws
 import arrow.typeclasses.Eq
 import arrow.typeclasses.EqK
